Item 내용들을 컴포넌트로 분리
